+++ packages/w3c-dtfm/src/alias.ts
@@ -3,7 +3,7 @@ import { TokenAlias } from './types';
 export function normalizeTokenAlias(tokenAlias: TokenAlias | string) {
   return (
     tokenAlias
-      // TODO: this is a hack to remove the $value from the token path, until we migrate isa figma plugin
+      // TODO: this is a hack to remove the $value from the token path, until we migrate isa's figma plugin
       .replace('.$value', '')
       .replace(/[{}]/g, '')
   );