+++ packages/w3c-dtfm/src/walker/index.ts
@@ -1,6 +1,7 @@
 export * from './extensions/generators';
 export * from './extensions/modifiers';
 export * from './extensions/responsive';
+export * from './token';
 export * from './types';
 export * from './utils';
 export * from './walker';