+++ apps/plugin/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @design-sync/figma-plugin
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @design-sync/utils@0.5.0
+  - @design-sync/rpc@0.0.6
+
 ## 0.0.4
 
 ### Patch Changes