+++ packages/w3c-dtfm/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @design-sync/w3c-dtfm
 
+## 0.10.0
+
+### Minor Changes
+
+- Improve the walker and make it lazy and add example apps
+
+### Patch Changes
+
+- Updated dependencies
+  - @design-sync/utils@0.5.0
+
 ## 0.9.1
 
 ### Patch Changes